melhorias no controller do people
